requirements.txt
flask
llama_index
torch
transformers
panda
numpy
langchain
python-dotenv
html2text